pyproject.toml
[tool.poetry]
name = "beanborg"
version = "0.9.0"
description = "Beanborg is a tool for automatically importing and categorizing financial transactions from CSV files into the Beancount bookkeeping system. It streamlines the process through account-matching rules, duplicate detection, and optional machine learning for transaction classification, making transaction imports quick and accurate."
authors = ["Luciano Fiandesio <luciano@fiandes.io>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/luciano-fiandesio/beanborg"
repository = "https://github.com/luciano-fiandesio/beanborg"


[tool.poetry.dependencies]
python = "^3.12"
beancount = "^3.0.0"
pyyaml = "^6.0.2"
rich = "^13.9.3"
prompt-toolkit = "^3.0.48"
pandas = "1.5.2"
textblob = "^0.18.0.post0"
openai = "^1.52.2"
numpy = "1.26.4"
fire = "^0.7.0"
imblearn = "^0.0"

[tool.poetry.scripts]
bb_mover = "beanborg.bb_mover:main"
bb_import = "beanborg.bb_import:main"
bb_archive = "beanborg.bb_archive:main"

[tool.poetry.group.dev.dependencies]
ruff = "^0.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "N",  # PEP8 naming convetions
    "D"  # pydocstyle
]
ignore = [
    "C901",  # too complex
    "W191",  # indentation contains tabs
    "D401"  # imperative mood
]

[tool.ruff.lint.pydocstyle]
convention = "google"